+++ src/main/java/com/codeit/deokhugam/service/GetPopularReviewsCommand.java
@@ -1,7 +1,6 @@
 package com.codeit.deokhugam.service;
 
 import com.codeit.deokhugam.domain.enums.Period;
-import java.math.BigDecimal;
 import java.time.Instant;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -15,7 +14,7 @@ public class GetPopularReviewsCommand {
 
   private final Period period;
   private final Direction direction;
-  private final BigDecimal cursor;
+  private final Long cursor;
   private final Instant after;
   private final Integer limit;
 }
